+++ src/store/setupStoreChannels.js
@@ -3,8 +3,7 @@ import log from 'loglevel'
 import pump from 'pump'
 import stream from 'stream'
 
-import { injectStore as onloadInjection } from '../onload'
-import torus from '../torus'
+import torus, { injectStore as onloadInjection } from '../torus'
 import { SUPPORTED_NETWORK_TYPES } from '../utils/enums'
 import { broadcastChannelOptions, isMain } from '../utils/utils'
 import { injectStore as controllerInjection } from './controllerSubscriptions'
